+++ experiment/HorizontalScrollBox_experiment/main.cpp
@@ -1,0 +1,10 @@
+#include "HorizontalScrollBox_experiment.h"
+#include <QtWidgets/QApplication>
+
+int main(int argc, char *argv[])
+{
+	QApplication a(argc, argv);
+	HorizontalScrollBox_experiment w;
+	w.show();
+	return a.exec();
+}